feat: improves home typing and update its unit tests
