+++ src/sections/home/types/homeTypes.ts
@@ -1,0 +1,12 @@
+export type HomeLabels = {
+  title: string
+  subtitle: string
+  description: string
+  primaryActionLabel: string
+  secondaryActionLabel: string
+}
+
+export type HomeActions = {
+  primaryAction: () => void
+  secondaryAction: () => void
+}
